Add Authentik setup in Ansible
This commit adds the Authentik setup in the Ansible config. It includes configuration for a new host under the group 'authentik' in the inventory, and a new ansible role 'authentik_host' to install Authentik using Docker with the necessary environment setup. It also adds a DNS record for 'authentik.campbell.lu'.